{"schema":"decodex/commit/1","summary":"{\"schema\":\"decodex/commit/1\",\"summary\":\"Split live adapter knowledge artifact modules\",\"authority\":\"manual\"}","authority":"manual"}
